# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ecommerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description?: string;
  imageUrl?: string;
  productCount: number;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  originalPrice?: number;
  imageUrl: string;
  images?: string[];
  categoryId: number;
  categoryName: string;
  stock: number;
  rating: number;
  reviewCount: number;
  isFeatured: boolean;
  isNew: boolean;
  badge?: string;
  tags?: string[];
}

export interface ProductsPage {
  products: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface StoreSummary {
  totalProducts: number;
  totalCategories: number;
  newArrivals: number;
  onSaleCount: number;
  topRatedCount: number;
}

export interface CartItem {
  productId: number;
  productName: string;
  price: number;
  originalPrice?: number;
  imageUrl: string;
  quantity: number;
  subtotal: number;
}

export interface Cart {
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface AddToCartBody {
  productId: number;
  quantity: number;
}

export interface UpdateCartItemBody {
  quantity: number;
}

export interface OrderItem {
  productId: number;
  productName: string;
  price: number;
  imageUrl: string;
  quantity: number;
  subtotal: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  items: OrderItem[];
  total: number;
  status: OrderStatus;
  customerName: string;
  customerEmail: string;
  shippingAddress: string;
  createdAt: string;
}

export interface CreateOrderBody {
  customerName: string;
  customerEmail: string;
  shippingAddress: string;
}

export type ListProductsParams = {
  category?: string;
  search?: string;
  page?: number;
  limit?: number;
  sort?: ListProductsSort;
};

export type ListProductsSort =
  (typeof ListProductsSort)[keyof typeof ListProductsSort];

export const ListProductsSort = {
  price_asc: "price_asc",
  price_desc: "price_desc",
  newest: "newest",
  popular: "popular",
} as const;