# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ecommerce API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: categories
    description: Category operations
  - name: cart
    description: Cart operations
  - name: orders
    description: Order operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 12
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [price_asc, price_desc, newest, popular]
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsPage"

  /products/featured:
    get:
      operationId: getFeaturedProducts
      tags: [products]
      summary: Get featured/hero products for homepage
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/summary:
    get:
      operationId: getProductsSummary
      tags: [products]
      summary: Get store summary stats
      responses:
        "200":
          description: Store summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSummary"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current cart
      responses:
        "200":
          description: Cart contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear the cart
      responses:
        "200":
          description: Cleared cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /cart/{productId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        productCount:
          type: integer
      required:
        - id
        - name
        - slug
        - productCount

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        stock:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        isFeatured:
          type: boolean
        isNew:
          type: boolean
        badge:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - description
        - price
        - imageUrl
        - categoryId
        - categoryName
        - stock
        - rating
        - reviewCount
        - isFeatured
        - isNew

    ProductsPage:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - limit
        - totalPages

    StoreSummary:
      type: object
      properties:
        totalProducts:
          type: integer
        totalCategories:
          type: integer
        newArrivals:
          type: integer
        onSaleCount:
          type: integer
        topRatedCount:
          type: integer
      required:
        - totalProducts
        - totalCategories
        - newArrivals
        - onSaleCount
        - topRatedCount

    CartItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        imageUrl:
          type: string
        quantity:
          type: integer
        subtotal:
          type: number
      required:
        - productId
        - productName
        - price
        - imageUrl
        - quantity
        - subtotal

    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        itemCount:
          type: integer
      required:
        - items
        - total
        - itemCount

    AddToCartBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required:
        - productId
        - quantity

    UpdateCartItemBody:
      type: object
      properties:
        quantity:
          type: integer
      required:
        - quantity

    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        quantity:
          type: integer
        subtotal:
          type: number
      required:
        - productId
        - productName
        - price
        - imageUrl
        - quantity
        - subtotal

    Order:
      type: object
      properties:
        id:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: number
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
        customerName:
          type: string
        customerEmail:
          type: string
        shippingAddress:
          type: string
        createdAt:
          type: string
      required:
        - id
        - items
        - total
        - status
        - customerName
        - customerEmail
        - shippingAddress
        - createdAt

    CreateOrderBody:
      type: object
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        shippingAddress:
          type: string
      required:
        - customerName
        - customerEmail
        - shippingAddress